pyproject.toml
[project]
name = "flood-assistance-system"
version = "0.1.0"
description = "Assistance system for users in vulnerable situations during flooding events, providing geospatial guidance, safety instructions and meteorological information."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.15.2",
    "langchain>=0.3.24",
    "langchain-google-genai>=2.1.3",
    "langchain-mcp-adapters>=0.0.9",
    "langgraph>=0.3.34",
    "mcp[cli]>=1.6.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.40",
    "flask (>=3.1.2,<4.0.0)",
    "flask-restful (>=0.3.10,<0.4.0)",
    "pytest>=8.4.2,<9.0.0",
]
